+++ doc/whats-new.rst
@@ -59,6 +59,9 @@ Documentation
 Internal Changes
 ~~~~~~~~~~~~~~~~
 
+- Removed deprecated method ``Dataset._replace_vars_and_dims`` and its 
+  internal usages (:pull:`3529`)
+  By `Maximilian Roos <https://github.com/max-sixty>`_
 
 - Removed internal method ``Dataset._from_vars_and_coord_names``, 
   which was dominated by ``Dataset._construct_direct``. (:pull:`3565`)